requirement.txt
asgiref==3.2.10
cassandra-driver==3.24.0
click==7.1.2
Django==3.1
django-cassandra-engine==1.6.1
djangorestframework==3.11.1
geomet==0.2.1.post1
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
